Added Styling
First draft of css. Added flexbox elements and finished the first draft of the score card. Colored and aligned elements of the Summary, struggling with adding gaps between them.
Everything used to be also aligned in the center of the page, but once I started working on Summary they all went to the top of the page.